Buff Jockey move target speed
If survivor is facing direction that jockey is not looking at, both players will stand still. This change will give full control of the player and Jockey will become a bit more similar to L4D with the ability to take the survivor away from the rest or to a deadly environment, bringing more tactics for this special infected.